Fit for 1920 x 1080 resolutions..
Finished the static design, need to implement api logic to get this
running. Images will be replaced with event pictures, map will be
populated with an actual google map.

This web application hasn't be developed as responsive yet, it'll be
tough to size the map accurately and place it under the listings.. but
that's a future plan.

I will add the uber and lyft buttons from the api, or I might just add
image hrefs with links leading to the website..

Finished.