Added request areas.
Might change lyft to 'Moov-it' for general use of public transportation.

Right now we have users only using two forms of expensive methods, I
believe it'd be more user-friendly if we incorporated a cheaper
alternative.

We also can just leave the two methods currently, and then, add the
alternative moov-it. But it may prove to be to much clutter.

We'll see.